backend/requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn[standard]==0.32.0
anthropic==0.40.0
pymupdf==1.24.14
python-pptx==1.0.2
python-multipart==0.0.12
pydantic==2.10.0
